test(frontend): move test dir to src
