+++ packages/frontend/src/__test__/fixtures/dummyApps.ts
